Move JVM-oriented functions to their own file
